tidied up the code for constants and functions
