+++ .eslintrc.js
@@ -1,11 +1,11 @@
 // https://docs.expo.dev/guides/using-eslint/
 module.exports = {
-  extends: ["expo", "eslint:recommended", "plugin:react-hooks/recommended", "prettier"],
+  extends: ["expo", "eslint:recommended", "plugin:react-hooks/recommended"],
   env: {
     "node": true,
     "react-native/react-native": true
   },
-  plugins: ['react-hooks', "react", "react-native", "react-refresh", "prettier"],
+  plugins: ['react-hooks', "react", "react-native", "react-refresh"],
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
@@ -16,6 +16,5 @@ module.exports = {
     "react-refresh/only-export-components": "warn",
     "react-hooks/rules-of-hooks": "error",
     "react-hooks/exhaustive-deps": "warn",
-    'prettier/prettier': 'error',
   }
 };
