+++ package.json
@@ -22,7 +22,7 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
-    "@gorhom/bottom-sheet": "^5.0.6",
+    "@gorhom/bottom-sheet": "^4.6.4",
     "@react-navigation/native": "^6.0.2",
     "eslint-config-expo": "^7.1.2",
     "eslint-plugin-react-native": "^4.1.0",
@@ -51,7 +51,7 @@
     "@types/jest": "^29.5.12",
     "@types/react": "~18.2.45",
     "@types/react-test-renderer": "^18.0.7",
-    "eslint-plugin-react-hooks": "^5.0.0",
+    "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.13",
     "jest": "^29.2.1",
     "jest-expo": "~51.0.3",
