+++ types/global.d.ts
@@ -1,8 +1,8 @@
-import { Difficulty } from "sudoku-gen/dist/types/difficulty.type";
+import { Difficulty } from "sudoku-gen/dist/types/difficulty.type"
 
 declare global {
     interface SudokuModeType {
-        id: string
-        difficulty: Difficulty
+        id: string;
+        difficulty: Difficulty;
     }
 }
